refactor(web): move `KeyboardObject` type to `common/web/types`
Fixes: #11612
