+++ web/src/app/browser/src/keymanEngine.ts
@@ -1,3 +1,4 @@
+import { KeyboardObject } from '@keymanapp/common-types';
 import { KeymanEngine as KeymanEngineBase, DeviceDetector } from 'keyman/engine/main';
 import { getAbsoluteY } from 'keyman/engine/dom-utils';
 import { OutputTarget } from 'keyman/engine/element-wrappers';
@@ -6,7 +7,7 @@ import {
   VisualKeyboard
 } from 'keyman/engine/osk';
 import { ErrorStub, KeyboardStub, CloudQueryResult, toPrefixedKeyboardId as prefixed } from 'keyman/engine/keyboard-storage';
-import { DeviceSpec, Keyboard, KeyboardObject } from "keyman/engine/keyboard";
+import { DeviceSpec, Keyboard } from "keyman/engine/keyboard";
 
 import * as views from './viewsAnchorpoint.js';
 import { BrowserConfiguration, BrowserInitOptionDefaults, BrowserInitOptionSpec } from './configuration.js';
