+++ web/src/engine/interfaces/src/prediction/languageProcessor.interface.ts
@@ -1,6 +1,5 @@
-import { Suggestion, Reversion } from '@keymanapp/common-types';
+import { OutputTarget, Suggestion, Reversion } from '@keymanapp/common-types';
 import { EventEmitter } from "eventemitter3";
-import { OutputTarget } from "keyman/engine/keyboard";
 
 export class ReadySuggestions {
   suggestions: Suggestion[];
